Add: simplified context usage
