+++ src/app/(deck-changers)/[slug]/view-import-deck/page.jsx
@@ -1,3 +1,4 @@
+import { useUserDeckContext } from '@/components/context-providers/userDeckContextProvider';
 import {
 	CommanderImage,
 	DeckInput,
@@ -12,28 +13,32 @@ export default async function ViewImportDeck({ searchParams }) {
      * Edit Mode: allows a user to edit and import the cards that are present in their deck. This is the default path if the user clicks NewDeck.
      * View Mode: Allows a user to view their cards and move onto suggestions
      */
-    
+    const { parseDeckInput, deparseDeckList,
+        verifyDeckList, printContext,
+        deckInput, setDeckInput,
+        deckList, setDeckList,
+        commander, deckName } = useUserDeckContext();
 
     return (
             <div className="flex flex-col items-center justify-center max-w-screen max-h-fill gap-8 p-4">
                 {/* First row: Deck Name and Button Group */}
                 <div className="flex flex-col-reverse w-full items-start gap-4 md:flex-row md:justify-between">
                     <div className="flex">
-                        <DeckNameInput /> 
+                        <DeckNameInput deckName={deckName} setDeckName={setDeckName} /> 
                     </div>
                     <div className="flex">
-                        <ViewButtonGroup />
+                        <ViewButtonGroup deckList={deckList}/>
                     </div>
                 </div>
                 <Divider />
                 {/* Second Row: Commander Selection */}
                 <div className="">
-                    <CommanderImage />
+                    <CommanderImage commander={commander} setCommander={setCommander}/>
                 </div>
                 <Divider />
                 {/* Third Row: Deck Input */}
                 <div className="w-full justify-start">
-                    <DeckInput />
+                    <DeckInput deckInput={deckInput} setDeckInput={setDeckInput}/>
                 </div>
             </div>
     );
